fix: 'last row problem' in b_range column
The auxiliary column `b_range` must always be one row longer than the main trace, in order to correctly verify range checks of the final value (u16::MAX). This is because the data used to generate the value included in the `b_range` column comes from the "current" row, but the value itself is included into the "next" row.

This was resolved by padding the main range checker trace with an extra row of the final value (u16::MAX). The 8-bit segment of the Range Checker's trace was adjusted accordingly with an additional lookup of ZERO (to account for the transition between the final two rows of the 16-bit segment of the trace where the value changes by 0).