docs(assembly): change examples for while, assertz
This commit updates the documentation for `while` and `assertz` of the
assembly documentation.

An example was added for `while` to clarify its expectation of having a
boolean on top of the stack, instead of using a constant `true` as
argument.

The `assertz` directive was updated from a nit that suggested there was
a `b` to be evaluated, when in fact it is a constant `0`.