chore(ci): add release to tests
This commit introduces the --release flag for the tests. It will
sacrifice some build time to greatly improve the tests execution.

One concern to add `release` is that `debug_assert` checks will no
longer be performed. These are used in `miden` and upstream
dependencies. To fix that, the `debug-assertions` compiler flag was
added so these assertions will be executed.

To ensure the tests always executes the debug assertions, a test
`debug_assert_is_checked` was added. It will expect a panic and
will fail by default.If it is set, then it will panic only if
`debug_assert` is evaluated.