refactor(processor): add range check trace table
This commit introduces `RangeCheckTraceTable`, a struct that will
contain the 8bit lookups table for the range trace, and the combined
length of the 8bits and 16bits tables.

Prior to this commit, the table was generated twice - first to calculate
the `max_len`, and then to compute the trace.

From local benchmarks, this optimization improved the performance around
~3%, considering we no longer allocate the tables twice. However, the
performance improvement could be mitigated by the compiler since he
should be smart enough to understand that the table shouldn't be dropped
before the second call to `into_trace`.

resolves #229