chore(ci): change clippy toolchain to nightly
This commit introduces night toolchain for clippy. The reason for that
is to stay up-to-date in `next`, and not merging changes that passes in
the CI but can break in the future with the stabilization of a clippy
lint.

If development is always done in nightly clippy, then any change that
lands from `next` to `main` is guaranteed to be compatible with future
stable versions of clippy.

As example, we have the PR #438 that was blocked by #447