refactor(assembly): library trait to provide ast modules
This commit changes the Library trait of assembly to provide AST modules
to the compilation process.

It also introduces a type-safe path for library namespace, module path
(relative), procedure name, and absolute path. Prior to this commit,
there was no distinction between these different states and they were
treated as raw strings. It was a context-heavy task to define when a
path is absolute and when is relative to a module.

The procedure ID, used to compile AST into MAST, is calculated always
with the absolute path of a procedure. Having only strings was
error-prone to compute the procedure ID correctly. With this commit, the
absolute path, used to compute a procedure ID, can only be obtained from
a sequence of validations in the namespace, module path, and procedure
name.

This commit also moves `ModuleProvider` to the private scope of the
assembly since now it is used only during the compilation context
creation and won't interact with the consumers of the library.

It will also fix a couple of new lints introduced in clippy for 1.67.

closes #552