refactor(assembly): add masl library serialization
This commit introduces the serialization of library contents into a MASL
(Miden Assembly Library) file. It will contain the root namespace,
version, and a list of modules that will be available to the
compilation.

This commit exposes the `Serializable` and `Deserializable` traits from
the parsers module to be used as standard for the
serialization/deserialization strategies. It also simplifies the build
process of the stdlib, creating a new instance on-demand via static
bytes included during the build process.

closes #469