Remove generated file
The file `stdlib/assets/std.masl` gets generated as part of the build process, but it is checked into the repository.

In general, it's a bad idea to check in generated files, as they can easily get out of sync and can easily be re-generated.

Fixes https://github.com/0xPolygonMiden/miden-vm/issues/620