feat(processor): advice provider as exec argument
This commit introduces the advice trait as argument of the execute
function of the processor. It will, downstream, impact on the prover and
miden binaries.

Currently, there is no mechanism to serialize a generic instance of the
advice provider to a file so it can be reconstructed for CLI
applications. The only use-case is to have in-memory, default
implementation of the advice provider, so it is using hardcoded
`MemAdviceProvider`.

The examples are also falling back to in-memory provider so we don't add
complexity and require them to serialize a generic advice provider.

closes #409