refactor(core): program inputs into advice inputs
Prior to this commit, the stack inputs were decoupled from the program
inputs so the advice provider can be generalized as a secret,
non-deterministic values/sets yielder. With that, the program inputs
became an ad hoc struct for the initialization of advice providers.

This commit moves the program inputs into the processor crate, so it can
be defined together with its only consumer, and renames it into
`AdviceInputs`.

It also renames `ProgramOutputs` to `StackOutputs` to maintain
consistency with the inputs.

This commit also activates documentation examples test for Miden and
adds the README as part of the library documentation. This will assert
the correctness of the examples in accordance with the new types.

related issue: #409