feat(processor): add program info
This commit introduces `ProgramInfo`, a structure that will bind both
the program hash and the used kernel to the generated zk proof.

It will use the kernel procedures as public inputs of the proof so the
security guarantees provided by the kernel can be extended to any state
transition with a valid proof. It will also allow efficient queries for
members of the kernel.

The CLI currently doesn't support kernel as argument as it will only
take the code root hash of a given program. It is currently using as
fallback and empty/default kernel.

related issue #409