change: remove unecessary column from runtime
It is not necessary to track the changes to the `row_addr` column during
the runtime, it is sufficient to construct the column when the trace is
being filled. This is because the value of the column can be inferred
from the any of the other columns in the hasher.

The advantage here is that the HasherTrace is slightly simpler, one
fewer attribute. And the column can be allocated and filled with a
single loop, which should be more efficient