feat: implement Deserializable for ProgramInfo
Currently, `ProgramInfo` implements only `Serializable`. It is so to
keep the coupling between winter-utils and miden minimal, as the
serialization traits are defined in winterfell.

However, this might create some challenges for consumers of miden-vm as
they might have to deserialize the struct, such as when representing the
type within other targets (i.e. WASM).

This commit introduces Deserializable for Kernel and ProgramInfo.

related issue: #731