chore: update cargo toml and makefile parameters
This commit removes the mandatory optimization from cargo toml and
creates makefile tasks to replace them.

They will allow the user to choose when he wants a full optimization, or
a partial. The full optimization is useful for release artifacts, and
partial is useful for development tasks such as tests.

closes #768