feat: replace merkle set by crypto store
A `MerkleStore` struct was introduced in `miden-crypto` that supersede
the `MerkleSet`. The advice provider must be refactored to use this new
structure.

The `MerkleStore` doesn't support the merkle copy in place, so the
mtree_cwn should be deprecated, and mtree_set updated to always copy.
This will be done in a subsequent PR.

closes #774