feat: deprecate mtree_cwm in favor of new mtree_merge
With the introduction of [MerkleStore], the copy in place for Merkle
trees became obsolete. This commit removes the `mtree_cwm` from the
instruction set.

This commit introduces a new instruction: `mtree_merge`. It will merge
the roots of two existing trees, increment their depth by 1 and creating
a new tree.

closes #775