refactor: advice provider merkle to return `Option`
Currently, the advice provider functions to query a Merkle store returns
a value wrapped by a `Result`. It means a "node not found" will be
treated as storage error.

However, this is not accurate; an absent node is not a storage failure,
and it might convey meaning under certain contexts.

One such example is the Sparse Merkle tree: an absent node means a leaf
is inserted on a tier smaller than the max bottom. This is a valid
state, and currently it is impossible to be represented by the advice
provider.

The current implementation is a reflection of how `miden-crypto` is
implemented. This PR intercepts the [MerkleError::NodeNotInStore] error
and returns `None`. However, ideally, [MerkleStore] should return `None`
for such cases.