feat: add `mtree_verify`
This commit introduces `mtree_verify`, a function that will perform a
raw [Operation::MpVerify]. It will not manipulate or consume elements of
the stack, and will be optimal in terms of efficiency - consuming a
single cycle.

It aims to support algorithms for multiple tree implementations - such
as Tiered Sparse Merkle tree - that will need to check for Merkle
openings while preserving the stack elements.

closes #848