feat: add line number to parsing error
Prior to this commit, [ParsingError] tracked the step of the source
instead of the line number.

This is not friendly to the user as he won't be able to find the
location of the error. A common standard is to display the line number,
so the user can easily spot the location of the error.

This commit introduces such track. It keeps a 1:1 map on [TokenStream]
from the token to the parsed line. As in common text editors, line
number starts at `1`.

It also removes the `pos` from [Token] as its only purpose was to
display the step to the user. The introduction of `line` completely
supersedes that.

closes #574