chore: enable resolver 2 for workspace
The resolver 2 needs to be explicitly enabled on workspaces [1]. Without
the `workspace.resolver=2` the old resolver 1 is used. Which during
feature unification will use unify with the bin and dev-dependencies,
even for release builds. The consequence of that is that running `cargo
build --no-default-features` on the workspace have most of the features
enabled because of the dev dependencies.

1- https://doc.rust-lang.org/cargo/reference/resolver.html#resolver-versions