refactor(assembly): improve ergonomics of and access to module import info
This commit changes the `import_info` fields of ProgramAst and ModuleAst
to remove the `Option` wrapper, since it does not appear to serve a
specific purpose, and both complicates access to the imports, and
requires redundant code in a few places to access common information
from the imports. It also required fallible checks in a few places where
empty module import info would be suitable as a fallback.

After this change, one can access the ModuleImports struct directly on
both ProgramAst and ModuleAst via the `import_info` function, which
returns a reference to the field. Redundant functions in both structs
were removed if they are already provided by ModuleImports.