feat: introduce thiserror for defining error types
This commit adds `thiserror` to the workspace for use in defining error
types, however it is a bit odd due to some background context you will
want to know:

Currently, the `std::error::Error` trait is only stable in libstd, and
its use in libcore is unstable, behind the `error_in_core` trait. This
makes defining `Error` types very awkward in general. As a result of
this, `thiserror` currently depends on libstd.

However, the crate author, dtolnay, has expressed that once
`error_in_core` stabilizes, `thiserror` will support no-std
environments. It is expected that this will stabilize soon-ish, but
there are no definite dates on feature stabilization.

Even though `thiserror` ostensibly requires libstd, it actually is
trivial to support no-std with it, by simply choosing _not_ to emit the
`std::error::Error` trait implementation when building for no-std, or
by enabling the `error_in_core` feature when building with nightly. The
crate author, dtolnay, has expressed that they would rather not maintain
that as a build option since `error_in_core` is so close to stabilization.
So to bridge the gap, I've forked `thiserror` and implemented the
necessary modifications, as well as `miette`, used for diagnostics,
which depends on `thiserror` internally.

In the future, when `thiserror` natively supports no-std builds, we can remove
these forked dependencies in favor of mainline.  In the meantime, we can
benefit from the use of `thiserror`'s ergonomic improvements when it comes
to defining error types, and it allows us to use `miette` as well.