wip(1/n): refactor procedure types
This commit is the first in a sequence of commits that represent the
refactoring of the `assembly` crate to use the new parser, etc., to
introduce the remaining AST changes, and then propagate those changes in
addition to refactoring parts of the compilation pipeline that can take
advantage of new features and analysis that were previously not
available.

This commit specifically refactors/rewrites the set of types which
represent various details about procedures and procedure "aliases", i.e.
re-exported procedures. There are some new types implemented as well, to
better represent the specificity of a particular procedure identifier,
and to build on other types in a more structured fashion.

NOTE: This commit references things which are not yet implemented in the
source tree, this is intentional so as to let you focus on this set of
related changes abstractly, and then be able to review other later
changes with this context in mind.