wip(3/n): refactor library types and api
This commit refactors the `LibraryPath` and `LibraryNamespace` types to
build on the `Ident` type, support `Spanned`, and to provide better
building blocks for other types, such as `FullyQualifiedProcedureName`.

In addition, the structure of the `library` namespace is cleaned up, and
the `MaslLibrary` type has its internals rewritten to make use of the
new parsing infrastructure.

The `Library` trait was refactored to remove the associated `ModuleIter`
type, as many forms of iterators have "unnameable" types which cannot be
easily expressed when defining a trait implementation. Instead, we make
use of RPIT (return-position impl trait) to acheive the same goal, while
retaining the ability to impose some useful constraints on the type.