wip(8/n): implement semantic analysis
This commit provides the implementation of the `sema` module that was
referenced in previous commits. It implements the translation of a
`Vec<Form>` to a `Module`, and runs various checks and rewrites on the
module as it is built. These are expressed using the `VisitMut` trait,
in two passes: `ConstEvalVisitor` (that does constant expression
evaluation), and `VerifyInvokeTargets` (that visits all invocation
targets and ensures that they are contextually valid, i.e. we can't know
if they are actually valid targets until we do global inter-procedural
analysis, but we can know if they reference an imported module, that
there is a corresponding import, etc.).

When semantic analysis is complete, a `Module` is locally valid, and
ready for compilation. The subsequent changes to the assembler will
demonstrate how we join together many modules, do global
inter-procedural analysis, and compile to MAST.