wip(9/n): implement module graph and call graph structures
This commit introduces two new data structures used in the rewrite of
the assembler's compilation pipeline: `ModuleGraph` and its twin, the
`CallGraph`.

The `ModuleGraph` represents the set of `Module`s being compiled as a
graph, where edges in the graph are formed by calls between modules.
These calls are unaliased, meaning that the edges refer to the concrete
module that an invoked procedure is defined in, not the module from
which it is re-exported (if a re-export is involved at all).

The graph is used to determine if we have any cyclic dependencies, i.e.
recursion in the graph which is not broken by either a dynamic call, or
by calling via MAST root (implying the called procedure does not depend
on any of the modules being compiled, or the hash would change after
compilation). In addition, this lets us do dead code analysis, and other
forms of inter-procedural analysis that would otherwise be difficult to
do without such a structure.

The `CallGraph` is used internally by `ModuleGraph` to represent the
actual graph structure itself, while `ModuleGraph` provides all of the
higher-level metadata and APIs, the `CallGraph` represents the graph
succinctly using a globally-unique identifiers for each procedure that
are stable within a `ModuleGraph`, but not usable across `ModuleGraph`s
(unless the other graph is a clone of the first).

`ModuleGraph`s are immutable once constructed, in that you cannot add
additional modules/libraries to them without building a new graph using
the `ModuleGraphBuilder` type. This is because a `ModuleGraph`
represents a valid compilation graph, and arbitrary changes could
violate those semantics, so we force them to be done via the builder to
enforce those properties in a single place.

That said, changes to a subset of the data in a `ModuleGraph` _is_
allowed, but only in specific ways.