wip(10/n): refactor procedure cache
This commit reimplements the procedure cache in a more efficient manner,
with the assumption that it is paired with a `ModuleGraph` in the
assmbler. Thus it can make various assumptions that we would otherwise be
unable to make.

It also provides a pretty-printed debug output that makes it easier to
review at a glance.