wip(11/10): refactor assembler interface
This commit reimplements the `Assembler` and `AssemblerContext` types to
be implemented in terms of the new module graph and procedure cache data
structures, with a slightly revamped, but more expansive interface.

The Assembler largerly still works as before, except it works
differently internally, and uses the AssemblyContext differently than
before. I'm still not 100% happy with having both types, but this PR
attempts to largely preserve the intent behind the AssemblyContext type.

Internally, the assembler builds up a set of modules to be compiled, and
then when `compile` (or one of its analogues) is invoked, the full set
of modules is analyzed and compiled, making use of previously built
items in the procedure cache when possible.

The primary visible difference with the new assembler is that, due to the
module graph, we are able to know the concrete MAST roots for all
callees in the call graph which are available on hand during
compilation. As a result, rather than inlining code blocks whenever an
`exec` is encountered, we make use of proxy blocks, which vastly reduces
the size of the MAST emitted. Support was added to the processor to
handle proxy blocks, since it appears that prior to this PR, proxy
blocks were treated as errors universally.

The new assembler also makes use of the new diagnostics system, and the
tests now reflect the diagnostics shown to users when an error occurs.