feat(assembler): add language support for explicit nop
This commit implements support in the parser and assembler of Miden
Assembly for the `nop` instruction, corresponding to the `Noop` opcode
in the VM.

This permits expressing certain patterns that can occur in generated
code, that while unlikely to occur in hand-written code, must
nevertheless be supported.

For example, when compiling conditional branches to MASM, when only one
of the branches actually does anything, cannot currently be represented
in MASM source code, e.g.:

```
if.true
else
    ..do stuff
end
```

In this example, some arbitrarily complex computation occurs only when
the condition is false, and the truthy branch is effectively empty.
Currently, `if.true` branches cannot be empty, and without the ability
to explicitly use `nop`, one must emit a sequence of useless
instructions so as to make the block non-empty.

NOTE: It is not sufficient to try and convert syntax like this into code
which always performs the computation of the non-empty branch, and then
drops the result, as the code in that branch may contain side effects.
Such optimization can be performed when it is deemed safe, and only
then.

Adding support for `nop` provides one solution to this problem. It also
provides us a way to print syntax trees with empty blocks in a canonical
form, so that transformations on the AST can be emitted in text form,
and correctly parsed. The compiler in particular would benefit from
this, as this is how it emits Miden Assembly today.

Eventually, the assembler should allow empty blocks in the parser, and
simply expand that to a block containing a single `nop` instruction, but
that is a convenience - `nop` is still the correct solution here.