feat: support absolute module/procedure paths
This commit does two things:

* Adds syntax for specifying absolute paths to procedures/modules in
  MASM, specifically, prefixing a path with `::` (just like in Rust)
  will treat the resulting path as absolute and fully-qualified,
  regardless of imports or other context.
* Ensures absolute paths are handled correctly throughout the assembler.
  Some support already existed, but was not complete due to lack of
  frontend implementation.

When absolute paths are used, one does not need to explicitly import the
module containing the referenced item, as there is no need for an import
when the dependency is explicit.

Due to uniformity of syntax representation in the parser, it is now
possible to define procedure aliases for items referenced by MAST root.
However, unless the module containing that root is provided during
compilation (or the procedure is available in the procedure cache), then
assembly will fail on these items. Additional refactoring will be
required to support phantom references like this fully, as the lack of
support at this time is due to the original assumptions of the
assembler, i.e. that all code will be available during assembly.