Remove `Library` trait (#1422)
* feat(assembler): implement serialization for `CompiledLibrary` and `KernelLibrary`
* feat(assembly): Add `CompiledLibrary::write_to_dir()` and `KernelLibrary::write_to_dir()`
* feat(assembler): Remove `MaslLibrary`
* remove docs building
* fix(assembly): ensure aliases are resolved to roots

This commit fixes the issue where re-exported procedures (aliases) were
not being handled properly during assembly. It now explicitly represents
them in the call graph, so that topological ordering of the call graph
will ensure that we only visit the aliases once the aliased procedure
has been compiled/visited.

As part of the solution implemented here, some refinements to
`AliasTarget` were made, in order to explicitly represent whether the
target is absolute or not (just like `InvocationTarget`). Additionally,
to avoid confusion, `FullyQualifiedProcedureName` was renamed to
`QualifiedProcedureName`, so as to make it clear that just because the
path is qualified, does not mean it is absolute (and thus "fully"
qualified).

Some conveniences were added to `LibraryNamespace`, `LibraryPath`, and
`AliasTarget` to make certain operations/conversions more ergonomic and
expressive.

* feat: add pretty-print helper for lists of values

* feat: support assembling with compiled libraries

This commit refactors `CompiledLibrary` a bit, to remove some
unnecessary restrictions leftover from the old MASL libraries:

* A `CompiledLibrary` no longer has a name, but it has a content digest
  obtained by lexicographically ordering the exported MAST roots of the
  library, and merging the hashes in order.
* As a consequence of being unnamed/having no namespace, a
  `CompiledLibrary` can now consist of procedures from many modules
  _and_ many namespaces. Any limitation we impose on top of that can be
  done via wrapper types, like how `KernelLibrary` is implemented.
* Support for re-exported procedures in a `CompiledLibrary` is
  implemented. It is assumed that all required libraries will be
  provided to the `Host` when executing a program.
* Some ergonomic improvements to APIs which accept libraries or sets of
  modules, to allow a greater variety of ways you can pass them.

* fix(assembly): address conflicting procedure definitions bug

Previously, we assumed that two procedures with the same MAST root, but
differing numbers of locals, was a bug in the assembler. However, this
is not the case, as I will elaborate on below.

If you compile a program like so:

```masm
use.std::u64

begin
  exec.u64::checked_and
end
```

The resulting MAST would look something like:

```mast
begin
  external.0x....
end
```

This MAST will have the exact same MAST root as `std::u64::checked_and`,
because `external` nodes have the same digest as the node they refer to.
Now, if the exec'd procedure has the same number of locals as the
caller, this is presumed to be a "compatible" procedure, meaning it is
fine to let both procedures refer to the same MAST. However, when the
number of procedure locals _differs_, we were raising a compatible
definition error, because it was assumed that due to the instructions
added when procedure locals are present, two procedures with differing
numbers of locals _could not_ have the same root by definition. This is
incorrect, let me illustrate:

```masm

export.foo.2
  ...
end

use.foo

proc.bar.3
   exec.foo::foo
end

begin
   exec.foo::foo
end
```

Assume that `foo.masm` is first compiled to a `CompiledLibrary`, which
is then added to the assembler when assembling an executable program
from `bar.masm`.

Also, recall that `begin .. end` blocks are effectively an exported
procedure definition, with zero locals, that has a special name - but in
all other ways it is just a normal procedure.

The above program is perfectly legal, but we would raise an error during
assembly of the program due to the `begin .. end` block compiling to an
`external` node referencing the MAST root of `foo::foo`, which has a
non-zero number of procedure locals.

The `bar` procedure is there to illustrate that even though it too
simply "wraps" the `foo::foo` procedure, it has a non-zero number of
procedure locals, and thus cannot ever have the same MAST root as a
wrapped procedure with a non-zero number of locals, due to the presence
of locals changing the root of the wrapping procedure. A check has been
kept around that ensures we catch if ever there are two procedures with
non-zero counts of procedure locals, with the same MAST root.

* chore: update changelog

---------

Co-authored-by: Bobbin Threadbare <bobbinth@protonmail.com>
Co-authored-by: Paul Schoenfelder <paulschoenfelder@fastmail.com>