feat(assembly): track source locations in debug mode (#1419)
* feat(assembly): track source locations in debug mode

This commit introduces an initial, minimal implementation of source
mapping for Miden Assembly instructions assembled to MAST. This is done
by piggy-backing on the existing `AssemblyOp` decorator, extending it
with an optional `SourceLocation`, which is only emitted to the
serialized MAST when debug mode is disabled _and_ there is source
location information available.

A `SourceLocation` consists of a string representing the file path of
the source (as known to the assembler), and a span, whose start and end
bounds are byte indices in the original source file, representing a
contiguous range of valid UTF-8 bytes corresponding to the relevant
source code from which the given Miden Assembly operation was derived.
This will be pretty much 1:1 when the source is itself MASM code, but
when the MASM in question is lowered from a high-level language, such as
Rust, the source span may correspond to many instructions, and the
connection between a given span and a specific MASM instruction may be
non-obvious at times - this is to be expected though, and is simply a
result of translating a high-level language to a low-level one.

The serialization used here is not particularly efficient, but isn't
horrible. In the future though, we may want to have a dedicated table
representation for source locations, so that streams of instructions
with a shared location can simply reference a single instance of the
source location, rather than duplicating the information for every
instruction. Additionally, we may want to allow including the actual
source file content in a library containing debug information, rather
than requiring end users to supply a path to a directory containing the
correct files.

One issue to be aware of, is that this implementation uses absolute
paths, but we may want to make paths relative when possible, so as to
make the paths portable. Technically, this is in the hands of users of
the assembler, by setting the source file paths to relative paths - but
in practice it would be quite awkard to force end users to handle this.
Additionally, there are times where an absolute path may be desirable
(such as when referencing well-known global paths, e.g.
`/usr/local/etc/..`). In any case, that is not a problem we attempt to
solve in this commit.

* Remove `SingleThreadedSourceManager`
* Rename `MultiThreadedSourceManager` to `DefaultSourceManager`
* Replace all uses of `SingleThreadedSourceManager` with
  `DefaultSourceManager`, remove all clippy attributes used to disable
  the lint about wrapping non-Send objects in an Arc
* Give `ProgramFile` two constructors, one for the common case which
  constructs a default source manager, the other for specific use cases
  where the caller owns the source manager and needs to construct the
  `ProgramFile` with it