cleanup: remove `miden example` command line option
It is redundant with `miden prove`, and runs code that is baked in to the CLI binary. This is not very useful to users, since they won't know the code that the example runs. It is better to just have people use `miden run` or `miden prove` on our example MASM programs.