fix(assembler): unnecessarily strict identifier validation
Ident, ProcedureName, LibraryNamespace, and LibraryPath, all support a
much less strict set of syntax for valid identifiers than Miden Assembly
source code does (at least, without quoting). Unfortunately, the default
constructor for these types, which was originally consumed only by the
parser, was performing validation on behalf of the parser that was in
fact unnecessary (as the lexer already does the bulk of this). This was
causing bugs in downstream projects, such as the compiler, which produce
identifiers that don't need to adhere to MASM source code identifier
rules because they lower directly to the AST.

This reworks the rules for identifiers allowed in these types, and
relies on the parser/lexer for the stricter validation where needed.

Closes 0xPolygonMiden/compiler#441