fix(assembly): invalid syntax for special identifiers
The names given to the special identifiers for executable, kernel, and
anonymous modules used `#` as a prefix, e.g. `#exec`, which is not valid
MASM syntax, as `#` is used for comments, and thus anything trailing
them is ignored on the same line.

This commit changes the prefix to `$`, and changes the name of the
kernel namespace from `#sys` to `$kernel` as well, to be clearer what
the identifier corresponds to.